+++ iot-sim/src/main/kotlin/Link.kt
@@ -1,7 +1,7 @@
 data class Link(
     val srcAddress: String,
     val destAddress: String,
-    var distanceInMeters: Double,
+    var distanceInMeters: Int,
     val linkType: LinkType
 )
